+++ app/api/router.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter
 
+from app.api.announcement import router as announcement_router
 from app.api.auth import router as auth_router
 from app.api.event import router as event_router
 from app.api.game_record import router as game_record_router
@@ -7,6 +8,7 @@ from app.api.group import router as group_router
 from app.api.user import router as user_router
 
 router = APIRouter()
+router.include_router(announcement_router)
 router.include_router(auth_router)
 router.include_router(group_router)
 router.include_router(game_record_router)